+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Support configuration of the slide buffer via lua functions
 - Support lua functions to for keymaps
 - Allow to disable existing keymaps by setting them to `nil`
+- Support configuration of the footer text
 
 ### Changed
 - Disable "release" workflow